Moved sqs header classes
