+++ instrumentation/aws-java-sdk-sqs-2.1.0/src/test/java/software/amazon/awssdk/services/sqs/SqsClientTest.java
@@ -35,7 +35,7 @@ import java.util.Map;
 import static org.junit.Assert.assertEquals;
 
 @RunWith(InstrumentationTestRunner.class)
-@InstrumentationTestConfig(includePrefixes = { "software.amazon.awssdk.services.sqs" }, configName = "dt_enabled.yml")
+@InstrumentationTestConfig(includePrefixes = { "software.amazon.awssdk.services.sqs", "com.newrelic.utils" }, configName = "dt_enabled.yml")
 public class SqsClientTest {
 
     private SqsClient sqsClient;
